fixes: - photo capture always being one frame behind - didn't cancel awaits on failed events - refactored callback since il2cpp doesn't support generics anyway

+++ CameraCapture/CameraCapture/Source/Shared/pch.h
@@ -37,7 +37,7 @@ struct CustomBuffer : winrt::implements<CustomBuffer, winrt::Windows::Storage::S
 
     uint32_t Capacity() const
     {
-        return m_buffer.size();
+        return static_cast<uint32_t>(m_buffer.size());
     }
 
     uint32_t Length() const
